+++ lib/DimPow.h
@@ -26,11 +26,11 @@
 #include <stdlib.h>
 
 typedef struct DimPow {
-    int  numer;       ///< Numerator of the power. May be negative.
-    int  denom;       ///< Denominator of the power. Will always be positive.
-    bool contributes; ///< If this contributes to the unit's dimensionality. For example, "g/g"
-                      ///< doesn't contribute to dimensionality but one might want that in a string
-                      ///< representation.
+    short numer;       ///< Numerator of the power. May be negative.
+    short denom;       ///< Denominator of the power. Will always be positive.
+    bool  contributes; ///< If this contributes to the unit's dimensionality. For example, "g/g"
+                       ///< doesn't contribute to dimensionality but one might want that in a string
+                       ///< representation.
 } DimPow; ///< Exponent of a unit dimension
 
 #ifdef __cplusplus
@@ -46,8 +46,8 @@ extern "C" {
  */
 void dp_init(
         DimPow* dp,
-        int     numer,
-        int     denom);
+        short   numer,
+        short   denom);
 
 /**
  * Default initializes.
@@ -66,8 +66,8 @@ void dp_initDefault(DimPow* dp);
  * @see `dp_free()`
  */
 DimPow* dp_new(
-    const int numer,
-    const int denom);
+    const short numer,
+    const short denom);
 
 /**
  * Returns a new, default dimension power.
@@ -89,7 +89,7 @@ void dp_free(DimPow* dp);
  * @param[in] dimPow  The dimension power
  * @return            The numerator of the power
  */
-int dp_getNumerator(const DimPow* dimPow);
+short dp_getNumerator(const DimPow* dimPow);
 
 /**
  * Returns the denominator of a dimension power.
@@ -97,7 +97,7 @@ int dp_getNumerator(const DimPow* dimPow);
  * @param[in] dimPow  The dimension power
  * @return            The denominator of the power
  */
-int dp_getDenominator(const DimPow* dimPow);
+short dp_getDenominator(const DimPow* dimPow);
 
 /**
  * Indicates if this dimension power contributes to a unit's dimensionality.
@@ -139,8 +139,8 @@ bool dp_equal(
  */
 void dp_multiply(
         DimPow* dimPow,
-        int     numer,
-        int     denom);
+        short   numer,
+        short   denom);
 
 /**
  * Adds to a dimension power. This is called when a unit is multiplied by another unit.
@@ -151,8 +151,8 @@ void dp_multiply(
  */
 void dp_add(
         DimPow* dimPow,
-        int     numer,
-        int     denom);
+        short   numer,
+        short   denom);
 
 /**
  * Returns a string representation of a dimension power. Formats include
